Move some stuff from HackyAI to new AIUtils
